+++ src/clj_async_profiler/post_processing.clj
@@ -1,6 +1,8 @@
 (ns clj-async-profiler.post-processing
-  (:require [clojure.java.io :as io])
+  (:require [clojure.java.io :as io]
+            [clojure.string :as str])
   (:import clojure.lang.Compiler
+           java.io.BufferedReader
            java.util.HashMap))
 
 (defn- safe-subs [^String s, ^long start, ^long end]
@@ -45,7 +47,7 @@
                                       ;; no special characters munged in the
                                       ;; frame, take a faster path.
                                       next-char-idx (inc uscore)]
-                                  (if (and (<= next-char-idx (.length ^String new-frame))
+                                  (if (and (< next-char-idx (.length ^String new-frame))
                                            (Character/isUpperCase (.charAt frame next-char-idx)))
                                     (Compiler/demunge new-frame)
                                     (-> ^String new-frame
@@ -72,15 +74,22 @@
             (do (.append sb (.substring s frame-beg frame-end))
                 (recur (unchecked-inc-int frame-end)))))))))
 
+(defn remove-lambda-ids
+  "Transform that removes numeric IDs next to anonymous functionas and `eval`
+  blocks. This is necessary to render a useful diffgraph if the diffed runs were
+  launched from the REPL."
+  [^String s]
+  (str/replace s #"(eval|__)\d+" "$1"))
+
 (defn post-process-stacks
   "Perform post-processing of the profiling result with the given `transform`
   function and the default processors (demunging). Write the result to
-  `out-file`."
+  `out-file`. Return the total number of samples in the file."
   [stacks-file out-file transform]
   (let [acc (HashMap.)]
     (with-open [f (io/reader stacks-file)]
       (loop []
-        (when-let [line (.readLine ^java.io.BufferedReader f)]
+        (when-let [line (.readLine ^BufferedReader f)]
           (let [sep (.lastIndexOf line " ")
                 stack (subs line 0 sep)
                 count (Long. (subs line (inc sep)))
@@ -94,4 +103,42 @@
       (binding [*out* out]
         (run! (fn [[stack count]]
                 (println stack count))
-              acc)))))
+              acc))
+      (reduce + (vals acc)))))
+
+(defn generate-diff-file
+  "Create a renderable diff stacks file from two provided stacks files."
+  [in-stacks-file1 in-stacks-file2 out-diff-file options]
+  (let [stacks (HashMap.)
+        normalize? (:normalize? options true)
+        transform (:transform options identity)
+        process-file
+        (fn [file tuple-idx]
+          (with-open [f (io/reader file)]
+            (loop [total-count 0]
+              (if-let [line (.readLine ^BufferedReader f)]
+                (let [sep (.lastIndexOf line " ")
+                      stack (-> (subs line 0 sep)
+                                remove-lambda-ids
+                                demunge-java-clojure-frames
+                                transform)]
+                  (if stack
+                    (let [count (Long. (subs line (inc sep)))]
+                      (when-not (.containsKey stacks stack)
+                        (.put stacks stack (long-array 2)))
+                      (aset ^longs (.get stacks stack) tuple-idx count)
+                      (recur (+ total-count count)))
+                    (recur total-count)))
+                total-count))))
+        total-count1 (process-file in-stacks-file1 0)
+        total-count2 (process-file in-stacks-file2 1)
+        scale (if normalize?
+                (/ (float total-count1) total-count2)
+                1)]
+    (with-open [f (io/writer out-diff-file)]
+      (doseq [[stack ^longs tuple] stacks
+              :let [c1 (aget tuple 0)
+                    ;; Normalize
+                    c2 (Math/round (double (* (aget tuple 1) scale)))]]
+        (binding [*out* f]
+          (println stack c2 c1))))))